answer_generation/pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "answer-generation"
version = "0.1.0"
description = "ViDoRe v3 answer generation and LLM-as-judge evaluation"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.5.0",
    "httpx>=0.27.0",
    "loguru>=0.7.3",
    "numpy",
    "pandas>=2.0.0",
    "pillow",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "torch",
    "tqdm",
    "typer",
    "visidata>=3.3",
]

[tool.ruff]
line-length = 99
